implemented Result system: it's like Optional<T>, but with an additional state representing an exception.
It is designed for return values of functions that would otherwise need
to throw checked exceptions.

A refactoring an function to "return Result.eval(() -> {....});"
encapsulates any thrown exceptions into a legal return state that can be
processed by the caller.

Base functional methods like map and flatMap are also supplied